// speedyGo-ui/src/app/services/models/reservation-carpool.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ReservationCarpool {
  accepted?: boolean;
  arrivalLocation?: string;
  bookingDate?: string;
  departLocation?: string;
  id?: number;
  requestedPrice?: number;
  requestedSeats?: number;
  status?: string;
}